Separated tests to different files
