+++ Examples/MovieProject/MovieProject.Tests/MovieProject.IsolatedTests/ComponentTesting/GetUserHappyTests.cs
@@ -4,8 +4,6 @@ using System.Collections.Generic;
 using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
-using MovieProject.Logic.DTO.MovieProject.Logic.Proxy.DTO;
-using Newtonsoft.Json;
 using SystemTestingTools;
 using Xunit;
 
@@ -17,8 +15,7 @@ namespace IsolatedTests.ComponentTestings
     {
         private readonly TestServerFixture Fixture;
 
-        private static string GetUrl = "https://jsonplaceholder.typicode.com/users";
-        private static string SearchUrl = "https://jsonplaceholder.typicode.com/searchUsers";
+        private static string Url = "https://jsonplaceholder.typicode.com/users";
 
         public GetUserHappyTests(TestServerFixture fixture)
         {        
@@ -38,7 +35,7 @@ namespace IsolatedTests.ComponentTestings
                 dto[0].Name = "Changed in code";
             });
 
-            client.AppendHttpCallStub(HttpMethod.Get, new System.Uri(GetUrl), response);
+            client.AppendHttpCallStub(HttpMethod.Get, new System.Uri(Url), response);
 
             // act
             var httpResponse = await client.GetAsync("/api/users");
@@ -52,7 +49,7 @@ namespace IsolatedTests.ComponentTestings
                 // assert outgoing
                 var outgoingRequests = client.GetSessionOutgoingRequests();
                 outgoingRequests.Count.Should().Be(1);
-                outgoingRequests[0].GetEndpoint().Should().Be($"GET {GetUrl}");
+                outgoingRequests[0].GetEndpoint().Should().Be($"GET {Url}");
                 outgoingRequests[0].GetHeaderValue("Referer").Should().Be(MovieProject.Logic.Constants.Website);
 
                 // assert return
@@ -70,50 +67,6 @@ namespace IsolatedTests.ComponentTestings
                 list[7].Should().Be("Mrs. Dennis Schulist");
                 list[8].Should().Be("Nicholas Runolfsdottir V");
                 list[9].Should().Be("Patricia Lebsack");
-            }
-        }
-        
-        [Fact]
-        public async Task When_UserSearchesWithValidParameters_Then_ReturnListProperly()
-        {
-            // arrange
-            var complexParameter = new UserSearchModel
-            {
-                Username = "Bret",
-            };
-            var serialisedComplexParameters = JsonConvert.SerializeObject(complexParameter);
-            var expectedResponse = new List<string>()
-            {
-                "Leanne Graham"
-            };
-            
-            var client = Fixture.Server.CreateClient();
-            client.CreateSession();
-            var response = ResponseFactory.FromFiddlerLikeResponseFile($"{Fixture.StubsFolder}/UserApi/Real_Responses/Happy/200_SearchListUsers.txt");
-
-            client.AppendHttpCallStub(HttpMethod.Get, new System.Uri(@$"{SearchUrl}?userSearchModel={serialisedComplexParameters}"), response);
-
-            // act
-            var httpResponse = await client.GetAsync("/api/users");
-
-            using (new AssertionScope())
-            {
-                // assert logs
-                var logs = client.GetSessionLogs();
-                logs.Should().BeEmpty();
-
-                // assert outgoing
-                var outgoingRequests = client.GetSessionOutgoingRequests();
-                outgoingRequests.Count.Should().Be(1);
-                outgoingRequests[0].GetEndpoint().Should().Be($"GET {SearchUrl}");
-                outgoingRequests[0].GetHeaderValue("Referer").Should().Be(MovieProject.Logic.Constants.Website);
-
-                // assert return
-                httpResponse.StatusCode.Should().Be(HttpStatusCode.OK);
-
-                var list = await httpResponse.ReadJsonBody<List<string>>();
-                list.Count.Should().Be(1);
-                list.Should().BeEquivalentTo(expectedResponse);
             }
         }
     }
